requirements.txt
Flask
Flask-JWT-Extended
Flask-RESTful
Flask-SqlAlchemy
passlib
PyJWT
SQLAlchemy
Werkzeug